Split StatementHelpers into multiple files
